rename views with Nuxt-style syntax
